requirements.txt
flask
flask_restplus
pyzmq
pyjwt
Flask-PyMongo